# Inline assume to avoid performance regression