@zakarumych 2023-03-22
Huge thanks for this 🥇 

I've copied this `Vec` from `std` almost as-is, removing mostly only some unstable API.
However it was slower than expected even causing performance penalty with generally faster allocator type.