@nical 2023-03-22
Happy to help!

Just thought about comparing against `allocator-api2` with the `nightly` feature which I think is more fair than comparing against the non-generic algorithm. Here are the results:

```
test fill_events_01_logo                ... bench:      33,624 ns/iter (+/- 1,109)
test fill_events_02_logo_pre_flattened  ... bench:      18,500 ns/iter (+/- 264)
test fill_events_03_logo_with_tess      ... bench:      85,580 ns/iter (+/- 2,876)
test fill_tess_01_logo                  ... bench:      68,030 ns/iter (+/- 266)
test fill_tess_03_logo_no_intersections ... bench:      78,257 ns/iter (+/- 409)
test fill_tess_05_logo_no_curve         ... bench:      32,906 ns/iter (+/- 77)
test fill_tess_06_logo_with_ids         ... bench:      66,430 ns/iter (+/- 2,005)
```

In conclusion:
 - The stable polyfill is still a regression but not as large as I made it look like in my previous comment.
 - (Goin in a tangent) slapping a generic parameter all over a pile of code can have non-trivial perf impacts even if always instantiated with the value used by the non-generic version.

In fact (for completeness in this tangent), here is the result when using `&dyn Allocator` (non-generic):

```
test fill_events_01_logo                ... bench:      31,632 ns/iter (+/- 957)
test fill_events_02_logo_pre_flattened  ... bench:      16,858 ns/iter (+/- 215)
test fill_events_03_logo_with_tess      ... bench:      79,375 ns/iter (+/- 2,620)
test fill_tess_01_logo                  ... bench:      62,266 ns/iter (+/- 755)
test fill_tess_03_logo_no_intersections ... bench:      70,584 ns/iter (+/- 1,236)
test fill_tess_05_logo_no_curve         ... bench:      31,057 ns/iter (+/- 786)
test fill_tess_06_logo_with_ids         ... bench:      62,483 ns/iter (+/- 1,220)
```

Dynamic dispatch + a pair of extra pointers in each Vec/Box + the indirection of this pair of pointers is faster than static dispatch with generics (measured on a m1 mac).