@zakarumych 2023-03-22
With `nightly` feature enabled `allocator-api2` reexports everything from `core`, `alloc` and `std` instead of defining replacements.
Note that `alloc::vec::Vec` and `alloc::boxed::Box` are already generic over allocator type. It's just defaulted to `Global` and usage of allocator type parameter is unstable.
So when you use `allocator_api2::vec::Vec<T, Global>` with `nightly` feature enabled it should be exactly the same as using `alloc::vec::Vec<T, Global>` that should be exactly the same as using `alloc::vec::Vec<T>`.