@zakarumych 2023-03-23
Ah, I see.
I've experimented a bit and found out that:
1. Passing `&A` around may be significantly worse for ZSTs like `Global` than using `A` directly.
   For the experiment in my custom allocator I was materializing generic `A` using `core::mem::zeroed()` inside branch with `size_of::<A>() == 0` condition instead of using `&A`. Compiler optimized out the branch and performance improved in few tests.
2. Wrapper type that fixes some generic parameter may slightly improve performance.