@zakarumych 2023-03-23
For lyon you can compare performance without generic allocator using `alloc::vec::Vec<T>`, `allocator_api2::vec::Vec<T>`.
And then with generic allocator using `allocator_api2::vec::Vec<T, &'a dyn Allocator>` and use `Global` and custom allocators like `blink-alloc` to see if performance improves significantly.